+++ src/persistence.ts
@@ -9,6 +9,7 @@ export interface CreateResourceRequest {
     resourceType: string;
     resource: any;
     id?: string;
+    ttlInSeconds?: number; // archive timestamp in Unix epoch time format in seconds
 }
 
 export interface UpdateResourceRequest {
@@ -16,6 +17,7 @@ export interface UpdateResourceRequest {
     resourceType: string;
     resource: any;
     vid?: string; // used in version aware update
+    ttlInSeconds?: number; // archive timestamp in Unix epoch time format in seconds
 }
 
 export interface PatchResourceRequest {
@@ -23,6 +25,7 @@ export interface PatchResourceRequest {
     resourceType: string;
     resource: any;
     vid?: string; // used in version aware patch
+    ttlInSeconds?: number; // archive timestamp in Unix epoch time format in seconds
 }
 
 export interface ReadResourceRequest {